+++ headers/Truck.h
@@ -12,26 +12,26 @@ public:
    using Vehicle::Vehicle;
 /** @brief Getter.
     @returns Truck maximum speed*/
-   const float  getSpeedMax() const;
+   const double  getSpeedMax() const;
 /** @brief Getter.
     @returns Truck acceleration*/
-   const float  getAcceleration() const;
+   const double  getAcceleration() const;
 /** @brief Getter.
     @returns Truck width*/
-   const float  getWidth() const;
+   const double  getWidth() const;
 /** @brief Getter.
     @returns Truck height*/
-   const float  getHeight() const;
+   const double  getHeight() const;
 /** @brief Getter.
     @returns Truck color*/
-   const float* getColor() const;
+   const double* getColor() const;
 /** @brief Getter.
     @returns Truck type*/
-   const bool   is2Wheeler() const;
+   const bool    is2Wheeler() const;
 protected:
 
 private:
 /** Parameters common to all trucks. */
-   static const float speedMax, acceleration, width, height, color[3];
+   static const double speedMax, acceleration, width, height, color[3];
    static const bool type;
 };
