+++ src/Truck.cpp
@@ -3,30 +3,30 @@
 #include "../headers/Constants.h"
 #include "../headers/Truck.h"
 
-const float Truck::speedMax     = constants::speedMaxTruck;
-const float Truck::acceleration = constants::accelerationTruck;
-const float Truck::width        = constants::widthTruck;
-const float Truck::height       = constants::heightTruck;
-const float Truck::color[3] = {0.0, 0.0, 0.9}; // Dark blue
+const double Truck::speedMax     = constants::speedMaxTruck;
+const double Truck::acceleration = constants::accelerationTruck;
+const double Truck::width        = constants::widthTruck;
+const double Truck::height       = constants::heightTruck;
+const double Truck::color[3] = {0.0, 0.0, 0.9}; // Dark blue
 const bool  Truck::type = false;
 
-const float Truck::getSpeedMax() const {
+const double Truck::getSpeedMax() const {
    return speedMax;
 }
 
-const float Truck::getAcceleration() const {
+const double Truck::getAcceleration() const {
    return acceleration;
 }
 
-const float Truck::getWidth() const {
+const double Truck::getWidth() const {
    return width;
 }
 
-const float Truck::getHeight() const {
+const double Truck::getHeight() const {
    return height;
 }
 
-const float* Truck::getColor() const {
+const double* Truck::getColor() const {
    return color;
 }
